Integrated new Parse Wizard with H2O -- designed, develped and tested all reqired APIs in core product code, added new features to parse process, merged with master

+++ py/testdir_release/kill_hadoop_jobs.sh
@@ -14,7 +14,7 @@ do
     # ignore these kind of lines
     # 0 jobs currently running
     # JobId   State   StartTime   UserName    Priority    SchedulingInfo
-    if [[ ("$jobid" != "JobId") && ("$state" != "jobs") ]]
+    if [[ ("$jobid" != "JobId") && ("$state" != "jobs") && ("$jobid" != "Total") ]]
     then
         echo "hadoop job -kill $jobid"
         $REMOTE_SSH_USER "hadoop job -kill $jobid"
